pyproject.toml
[project]
name = "web-shop-fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-slugify>=8.0.4",
    "ruff>=0.14.11",
    "sqladmin>=0.22.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "mypy>=1.19.1",
    "pyjwt[crypto]>=2.11.0",
    "black>=26.1.0",
    "python-multipart>=0.0.22",
    "aiofiles>=25.1.0",
]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "SIM",
]

[tool.ruff.lint.isort]
combine-as-imports = true
